+++ packages/gits_base/lib/gits_base.dart
@@ -3,5 +3,4 @@ library gits_base;
 export 'package:flutter_bloc/flutter_bloc.dart';
 
 export 'src/gits_cubit.dart';
-export 'src/gits_hydrated.dart';
 export 'src/gits_state_page.dart';