+++ packages/gits_inspector/lib/src/helper/local_notification.dart
@@ -58,7 +58,7 @@ final class LocalNotification {
       await _flutterLocalNotificationsPlugin
           .resolvePlatformSpecificImplementation<
               AndroidFlutterLocalNotificationsPlugin>()
-          ?.requestPermission();
+          ?.requestNotificationsPermission();
     }
   }