+++ packages/gits_inspector/lib/src/widgets/theme_inspector.dart
@@ -3,9 +3,9 @@ import 'package:flutter/material.dart';
 /// Default theme inspector with dark mode.
 class ThemeInspector extends StatelessWidget {
   const ThemeInspector({
-    Key? key,
+    super.key,
     required this.child,
-  }) : super(key: key);
+  });
 
   final Widget child;