+++ packages/gits_responsive/lib/gits_responsive.dart
@@ -1,0 +1,6 @@
+export 'src/gits_breakpoint.dart';
+export 'src/gits_inherited_breakpoint.dart';
+export 'src/gits_responsive.dart';
+export 'src/gits_responsive_extension.dart';
+export 'src/gits_responsive_target.dart';
+export 'src/gits_responsive_widget.dart';