+++ packages/gits_responsive/lib/src/gits_responsive_target.dart
@@ -1,0 +1,6 @@
+/// The class defines constants for different types of responsive targets.
+abstract class GitsResponsiveTarget {
+  static const String mobile = "MOBILE";
+  static const String tablet = "TABLET";
+  static const String desktop = "DESKTOP";
+}