+++ packages/gits_cached_network_image/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.2
+
+- upgrade version http to 1.1.2
+- upgrade flutter_lint, hive_generator and build_runner to latest
+
 ## 1.0.1
 
 - upgrade version http to 1.1.0